CARNET-2; Make code uglier and faster; more precise test; <cargo>Make code uglier and faster; more precise test</cargo>
The function ignores non-string values. This has always been the case. No behavior change here.

Addtionally, the function's now consistently 4x faster due to less allocations. I kept the allocation for `classNames`, because it turns out that it's faster to mutate an array and join it at the end than to continuously append to a string, especially for slightly longer class names.